more efficient rectangle cost evaluation using Double instead of BigDecimal

+++ bin-packing-backend-play/app/models/algorithm/LocalSearch.scala
@@ -66,7 +66,7 @@ trait SolutionHandler[Solution] {
 trait Score extends Ordered[Score]
 
 case class OneDimensionalScore(
-  value: BigDecimal
+  value: Double
 ) extends Score {
   override def compare(that: Score): Int = value.compare(that.asInstanceOf[OneDimensionalScore].value)
 }
